+++ docs/guides/changelog.md
@@ -9,8 +9,8 @@ description: |-
 
 ## 0.2.17 (unreleased)
 
+- Add `ise_trustsec_work_process_settings` resource
 - Add `ise_trustsec_matrix` resource and data source
-
 
 ## 0.2.16